1. Fix the bug : when user select A subs in azure cli but use B subs in configuration 2. Fix the bug : don't prompt current subscription used in spring cloud plugin 3. Fix the bug : the subscription id in telemetry of spring cloud plugin 4. Refactor:  remove duplicate code in AbstractAzureMojo and AbstractMojoBase 5. Fix the bug: config will not choose subscription on azure cli
